Flytt kombinering-funksjoner til egen fil
