+++ tidslinje/src/main/kotlin/no/nav/familie/tidslinje/utvidelser/TidslinjeUtvidelser.kt
@@ -14,7 +14,6 @@ import no.nav.familie.tidslinje.filtrerIkkeNull
 import no.nav.familie.tidslinje.tilPeriodeVerdi
 import java.time.LocalDate
 import java.time.temporal.ChronoUnit
-import kotlin.math.absoluteValue
 
 val TIDENES_ENDE = LocalDate.MAX
 
@@ -482,31 +481,6 @@ private fun LocalDate.leggTil(
         TidsEnhet.MÅNED -> this.plusMonths(antall)
         TidsEnhet.ÅR -> this.plusYears(antall)
     }
-
-fun <T, R, RESULTAT> Tidslinje<T>.kombinerMed(
-    annen: Tidslinje<R>,
-    kombineringsfunksjon: (elem1: T?, elem2: R?) -> RESULTAT?,
-): Tidslinje<RESULTAT> =
-    this.biFunksjon(annen) { periodeverdiVenstre, periodeverdiHøyre ->
-        kombineringsfunksjon(periodeverdiVenstre.verdi, periodeverdiHøyre.verdi)
-            .tilPeriodeVerdi()
-    }
-
-fun <T, R, S, RESULTAT> Tidslinje<T>.kombinerMed(
-    tidslinje2: Tidslinje<R>,
-    tidslinje3: Tidslinje<S>,
-    kombineringsfunksjon: (elem1: T?, elem2: R?, elem3: S?) -> RESULTAT?,
-): Tidslinje<RESULTAT> {
-    val tidslinje1Og2: Tidslinje<Pair<T?, R?>> =
-        this.biFunksjon(tidslinje2) { elem1PeriodeVerdi, elem2PeriodeVerdi ->
-            Pair(elem1PeriodeVerdi.verdi, elem2PeriodeVerdi.verdi).tilPeriodeVerdi()
-        }
-
-    return tidslinje1Og2.biFunksjon(tidslinje3) { elem1Og2, elem3PeriodeVerdi ->
-        val (elem1, elem2) = elem1Og2.verdi ?: Pair(null, null)
-        kombineringsfunksjon(elem1, elem2, elem3PeriodeVerdi.verdi).tilPeriodeVerdi()
-    }
-}
 
 fun <T> Tidslinje<T>.tilPerioder(): List<Periode<T?>> = this.tilTidslinjePerioderMedDato().map { it.tilPeriode() }
